+++ lib/pages/livestock/feed-record/feed-livestock-management/add_livestock_feed_type.dart
@@ -113,7 +113,7 @@ class _AddLiveStockFeedTypeState extends State<AddLiveStockFeedType> {
                                         feedname: feedname.text,
                                         feedtype: feedtype.text,
                                         quantityaday:
-                                            int.parse(quantityaday.text),
+                                            double.parse(quantityaday.text),
                                         livestockname: livestockname.text,
                                         livestockid: livestockid.text,
                                         animalweight:
